Show resources in Slack notification for access requests
This change updates Slack notifications for resource-based access
requests to include the resources being requested.
