+++ integrations/access/common/message.go
@@ -78,8 +78,11 @@ func MsgFields(reqID string, reqData pd.AccessRequestData, clusterName string, w
 	if len(reqData.User) > 0 {
 		msgFieldToBuilder(&builder, "User", reqData.User)
 	}
-	if reqData.Roles != nil {
+	if len(reqData.Roles) > 0 {
 		msgFieldToBuilder(&builder, "Role(s)", strings.Join(reqData.Roles, ","))
+	}
+	if len(reqData.Resources) > 0 {
+		msgFieldToBuilder(&builder, "Resource(s)", strings.Join(reqData.Resources, ","))
 	}
 	if reqData.RequestReason != "" {
 		msgFieldToBuilder(&builder, "Reason", lib.MarkdownEscape(reqData.RequestReason, requestReasonLimit))
