+++ integrations/access/common/plugindata_test.go
@@ -26,6 +26,7 @@ var samplePluginData = GenericPluginData{
 	AccessRequestData: plugindata.AccessRequestData{
 		User:             "user-foo",
 		Roles:            []string{"role-foo", "role-bar"},
+		Resources:        []string{"cluster-a/node/foo", "cluster-a/node/bar"},
 		RequestReason:    "foo reason",
 		ReviewsCount:     3,
 		ResolutionTag:    plugindata.ResolvedApproved,
@@ -40,9 +41,10 @@ var samplePluginData = GenericPluginData{
 func TestEncodePluginData(t *testing.T) {
 	dataMap, err := EncodePluginData(samplePluginData)
 	assert.NoError(t, err)
-	assert.Len(t, dataMap, 7)
+	assert.Len(t, dataMap, 8)
 	assert.Equal(t, "user-foo", dataMap["user"])
 	assert.Equal(t, "role-foo,role-bar", dataMap["roles"])
+	assert.Equal(t, `["cluster-a/node/foo","cluster-a/node/bar"]`, dataMap["resources"])
 	assert.Equal(t, "foo reason", dataMap["request_reason"])
 	assert.Equal(t, "3", dataMap["reviews_count"])
 	assert.Equal(t, "APPROVED", dataMap["resolution"])
@@ -54,6 +56,7 @@ func TestDecodePluginData(t *testing.T) {
 	pluginData, err := DecodePluginData(map[string]string{
 		"user":           "user-foo",
 		"roles":          "role-foo,role-bar",
+		"resources":      `["cluster-a/node/foo","cluster-a/node/bar"]`,
 		"request_reason": "foo reason",
 		"reviews_count":  "3",
 		"resolution":     "APPROVED",
@@ -67,7 +70,7 @@ func TestDecodePluginData(t *testing.T) {
 func TestEncodeEmptyPluginData(t *testing.T) {
 	dataMap, err := EncodePluginData(GenericPluginData{})
 	assert.NoError(t, err)
-	assert.Len(t, dataMap, 7)
+	assert.Len(t, dataMap, 8)
 	for key, value := range dataMap {
 		assert.Emptyf(t, value, "value at key %q must be empty", key)
 	}
