+++ integrations/lib/plugindata/access_request_test.go
@@ -23,6 +23,7 @@ import (
 var sampleAccessRequestData = AccessRequestData{
 	User:             "user-foo",
 	Roles:            []string{"role-foo", "role-bar"},
+	Resources:        []string{"cluster/node/foo", "cluster/node/bar"},
 	RequestReason:    "foo reason",
 	ReviewsCount:     3,
 	ResolutionTag:    ResolvedApproved,
@@ -32,9 +33,10 @@ var sampleAccessRequestData = AccessRequestData{
 func TestEncodeAccessRequestData(t *testing.T) {
 	dataMap, err := EncodeAccessRequestData(sampleAccessRequestData)
 	assert.Nil(t, err)
-	assert.Len(t, dataMap, 6)
+	assert.Len(t, dataMap, 7)
 	assert.Equal(t, "user-foo", dataMap["user"])
 	assert.Equal(t, "role-foo,role-bar", dataMap["roles"])
+	assert.Equal(t, `["cluster/node/foo","cluster/node/bar"]`, dataMap["resources"])
 	assert.Equal(t, "foo reason", dataMap["request_reason"])
 	assert.Equal(t, "3", dataMap["reviews_count"])
 	assert.Equal(t, "APPROVED", dataMap["resolution"])
@@ -45,6 +47,7 @@ func TestDecodeAccessRequestData(t *testing.T) {
 	pluginData, err := DecodeAccessRequestData(map[string]string{
 		"user":           "user-foo",
 		"roles":          "role-foo,role-bar",
+		"resources":      `["cluster/node/foo", "cluster/node/bar"]`,
 		"request_reason": "foo reason",
 		"reviews_count":  "3",
 		"resolution":     "APPROVED",
@@ -57,7 +60,7 @@ func TestDecodeAccessRequestData(t *testing.T) {
 func TestEncodeEmptyAccessRequestData(t *testing.T) {
 	dataMap, err := EncodeAccessRequestData(AccessRequestData{})
 	assert.Nil(t, err)
-	assert.Len(t, dataMap, 6)
+	assert.Len(t, dataMap, 7)
 	for key, value := range dataMap {
 		assert.Emptyf(t, value, "value at key %q must be empty", key)
 	}
