--- lib/services/friendlyname.go
@@ -1,34 +1,0 @@
-/*
-Copyright 2023 Gravitational, Inc.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package services
-
-import "github.com/gravitational/teleport/api/types"
-
-// FriendlyName will return the friendly name for a resource if it has one. Otherwise, it
-// will return an empty string.
-func FriendlyName(resource types.ResourceWithLabels) string {
-	// Right now, only resources sourced from Okta and nodes have friendly names.
-	if resource.Origin() == types.OriginOkta {
-		return resource.GetMetadata().Description
-	}
-
-	if hn, ok := resource.(interface{ GetHostname() string }); ok {
-		return hn.GetHostname()
-	}
-
-	return ""
-}
