+++ lib/web/ui/user_groups.go
@@ -18,7 +18,6 @@ package ui
 
 import (
 	"github.com/gravitational/teleport/api/types"
-	"github.com/gravitational/teleport/lib/services"
 )
 
 // UserGroup describes a user group.
@@ -54,7 +53,7 @@ func MakeUserGroups(userGroups []types.UserGroup, userGroupsToApps map[string]ty
 		for i, app := range apps {
 			appsAndFriendlyNames[i] = ApplicationAndFriendlyName{
 				Name:         app.GetName(),
-				FriendlyName: services.FriendlyName(app),
+				FriendlyName: types.FriendlyName(app),
 			}
 		}
 
@@ -62,7 +61,7 @@ func MakeUserGroups(userGroups []types.UserGroup, userGroupsToApps map[string]ty
 			Name:         userGroup.GetName(),
 			Description:  userGroup.GetMetadata().Description,
 			Labels:       uiLabels,
-			FriendlyName: services.FriendlyName(userGroup),
+			FriendlyName: types.FriendlyName(userGroup),
 			Applications: appsAndFriendlyNames,
 		})
 	}
